added event bus; moved some http network calls to temporary class - preparing to use retrofit; replaced ListView with RecyclerView in ListaVijestiActivity; creating advanced MenuItemAdapter; refactored ListaVijestiActivity to use Fragments when traversing directory tree instead of creating new activity instances; replaced ProgressDialog in ListaVijestiActivity with progress bar inside RecyclerView; lazily creating typefaces from file in unified place; other major and minor refactor and cleanup

+++ app/src/main/java/hr/bpervan/novaeva/utilities/Constants.java
@@ -69,7 +69,7 @@ public class Constants {
 	public static final String searchURL = "http://novaeva.com/json?api=2&search=";
 	public static final String alertURL = "http://novaeva.com/json?api=2&alert=1";
 	
-	public static String getCatNameById(int id){
+	public static String getCatNameById(long id){
 		if(id == CAT_AKTUALNO)
 			return CAT_AKTUALNO_NAZIV;
 		else if(id == CAT_EVANDJELJE)
